healthcheck/requirements.txt
connexion[flask,uvicorn,swagger-ui]
httpx
APScheduler
PyYAML
requests
python-json-logger